+++ linkerd/identifier/http/src/main/scala/io/l5d/HttpIdentifierConfig.scala
@@ -1,0 +1,19 @@
+package io.l5d
+
+import com.fasterxml.jackson.annotation.JsonAutoDetect.Visibility
+import com.fasterxml.jackson.annotation.{JsonAutoDetect, JsonIgnore, JsonTypeInfo}
+import com.twitter.finagle.{Dtab, Path, http}
+import io.buoyant.router.RoutingFactory.Identifier
+
+@JsonTypeInfo(use = JsonTypeInfo.Id.NAME, include = JsonTypeInfo.As.PROPERTY, property = "kind")
+@JsonAutoDetect(fieldVisibility = Visibility.ANY)
+trait HttpIdentifierConfig {
+  var httpUriInDst: Option[Boolean] = None
+
+  @JsonIgnore
+  def newIdentifier(
+    prefix: Path,
+    baseDtab: () => Dtab = () => Dtab.base
+  ): Identifier[http.Request]
+}
+